Conflicts:
	ps3mediaserver/contrib/binaries-deps-versions
	ps3mediaserver/contrib/download-pms-binaries-source.sh
	ps3mediaserver/contrib/mplayer-r34393-SB19.patch
	ps3mediaserver/pom.xml
	ps3mediaserver/src/main/java/net/pms/formats/FormatFactory.java
	ps3mediaserver/src/test/java/net/pms/test/formats/FormatFactoryTest.java
	ps3mediaserver/src/test/java/net/pms/test/formats/FormatRecognitionTest.java
	ps3mediaserver/src/test/java/net/pms/test/formats/FormatTest.java
	src/main/external-resources/logback.xml

+++ ps3mediaserver/src/main/java/net/pms/util/FlowParserOutputStream.java
@@ -35,7 +35,7 @@ public abstract class FlowParserOutputStream extends OutputStream {
 				buffer.put(swapRemainingByte[0]);
 				off++;
 			}
-			int modulo = (int) Math.abs(len - off) % swapOrderBits;
+			int modulo = Math.abs(len - off) % swapOrderBits;
 			if (modulo != 0) {
 				swapRemainingByte = new byte[1];
 				len -= modulo;
